+++ .changeset/olive-jeans-reply.md
@@ -1,0 +1,5 @@
+---
+'@astrojs/cloudflare': patch
+---
+
+Refactor codebase to enhance code readability and structure, to prioritize maintainability for long-term.
