+++ packages/integrations/cloudflare/src/entrypoints/server.advanced.ts
@@ -1,16 +1,20 @@
-import type { Request as CFRequest, EventContext } from '@cloudflare/workers-types';
+import type { Request as CFRequest, ExecutionContext } from '@cloudflare/workers-types';
 import type { SSRManifest } from 'astro';
 import { App } from 'astro/app';
-import { getProcessEnvProxy, isNode } from './util.js';
+import { getProcessEnvProxy, isNode } from '../util.js';
 
 if (!isNode) {
 	process.env = getProcessEnvProxy();
 }
 
-export interface DirectoryRuntime<T extends object = object> {
+type Env = {
+	ASSETS: { fetch: (req: Request) => Promise<Response> };
+};
+
+export interface AdvancedRuntime<T extends object = object> {
 	runtime: {
 		waitUntil: (promise: Promise<any>) => void;
-		env: EventContext<unknown, string, unknown>['env'] & T;
+		env: Env & T;
 		cf: CFRequest['cf'];
 		caches: typeof caches;
 	};
@@ -19,15 +23,13 @@ export interface DirectoryRuntime<T extends object = object> {
 export function createExports(manifest: SSRManifest) {
 	const app = new App(manifest);
 
-	const onRequest = async (context: EventContext<unknown, string, unknown>) => {
-		const request = context.request as CFRequest & Request;
-		const { env } = context;
-
+	const fetch = async (request: Request & CFRequest, env: Env, context: ExecutionContext) => {
 		// TODO: remove this any cast in the future
 		// REF: the type cast to any is needed because the Cloudflare Env Type is not assignable to type 'ProcessEnv'
 		process.env = env as any;
 
 		const { pathname } = new URL(request.url);
+
 		// static assets fallback, in case default _routes.json is not used
 		if (manifest.assets.has(pathname)) {
 			return env.ASSETS.fetch(request);
@@ -41,12 +43,12 @@ export function createExports(manifest: SSRManifest) {
 				request.headers.get('cf-connecting-ip')
 			);
 
-			const locals: DirectoryRuntime = {
+			const locals: AdvancedRuntime = {
 				runtime: {
 					waitUntil: (promise: Promise<any>) => {
 						context.waitUntil(promise);
 					},
-					env: context.env,
+					env: env,
 					cf: request.cf,
 					caches: caches,
 				},
@@ -69,5 +71,5 @@ export function createExports(manifest: SSRManifest) {
 		});
 	};
 
-	return { onRequest, manifest };
+	return { default: { fetch } };
 }
