+++ packages/integrations/cloudflare/src/entrypoints/server.directory.ts
@@ -1,20 +1,16 @@
-import type { Request as CFRequest, ExecutionContext } from '@cloudflare/workers-types';
+import type { Request as CFRequest, EventContext } from '@cloudflare/workers-types';
 import type { SSRManifest } from 'astro';
 import { App } from 'astro/app';
-import { getProcessEnvProxy, isNode } from './util.js';
+import { getProcessEnvProxy, isNode } from '../util.js';
 
 if (!isNode) {
 	process.env = getProcessEnvProxy();
 }
 
-type Env = {
-	ASSETS: { fetch: (req: Request) => Promise<Response> };
-};
-
-export interface AdvancedRuntime<T extends object = object> {
+export interface DirectoryRuntime<T extends object = object> {
 	runtime: {
 		waitUntil: (promise: Promise<any>) => void;
-		env: Env & T;
+		env: EventContext<unknown, string, unknown>['env'] & T;
 		cf: CFRequest['cf'];
 		caches: typeof caches;
 	};
@@ -23,13 +19,15 @@ export interface AdvancedRuntime<T extends object = object> {
 export function createExports(manifest: SSRManifest) {
 	const app = new App(manifest);
 
-	const fetch = async (request: Request & CFRequest, env: Env, context: ExecutionContext) => {
+	const onRequest = async (context: EventContext<unknown, string, unknown>) => {
+		const request = context.request as CFRequest & Request;
+		const { env } = context;
+
 		// TODO: remove this any cast in the future
 		// REF: the type cast to any is needed because the Cloudflare Env Type is not assignable to type 'ProcessEnv'
 		process.env = env as any;
 
 		const { pathname } = new URL(request.url);
-
 		// static assets fallback, in case default _routes.json is not used
 		if (manifest.assets.has(pathname)) {
 			return env.ASSETS.fetch(request);
@@ -43,12 +41,12 @@ export function createExports(manifest: SSRManifest) {
 				request.headers.get('cf-connecting-ip')
 			);
 
-			const locals: AdvancedRuntime = {
+			const locals: DirectoryRuntime = {
 				runtime: {
 					waitUntil: (promise: Promise<any>) => {
 						context.waitUntil(promise);
 					},
-					env: env,
+					env: context.env,
 					cf: request.cf,
 					caches: caches,
 				},
@@ -71,5 +69,5 @@ export function createExports(manifest: SSRManifest) {
 		});
 	};
 
-	return { default: { fetch } };
+	return { onRequest, manifest };
 }
