+++ packages/integrations/cloudflare/src/utils/prependForwardSlash.ts
@@ -1,0 +1,3 @@
+export function prependForwardSlash(path: string) {
+	return path[0] === '/' ? path : '/' + path;
+}
